+++ km3pipe/calib.py
@@ -27,7 +27,12 @@ log = get_logger(__name__)
 
 
 class Calibration(Module):
-    """A very simple, preliminary Module which gives access to the calibration.
+    """A module which applies time, position and rotation corrections to hits.
+
+    This module also calibrates MC hits, but be aware, t0s are not appended to
+    the MC hit times.
+    Additionally, the global PMT ID is added to regular hits as ``pmt_id`` and
+    in case of MC hits, the ``dom_id`` and ``channel_id`` (DAQ) are set.
 
     Parameters
     ----------
@@ -41,6 +46,14 @@ class Calibration(Module):
         t0set (when retrieving from database).
     calibset: optional
         calibset (when retrieving from database).
+    key: str, optional [default="Hits"]
+        the blob key of the hits
+    outkey: str, optional [default="CalibHits"]
+        the output blob key of the calibrated hits
+    key_mc: str, optional [default="McHits"]
+        the blob key of the MC hits (if present)
+    outkey_mc: str, optional [default="CalibMcHits"]
+        the output blob key of the calibrated MC hits
     """
 
     __name__ = "Calibration"
@@ -54,6 +67,10 @@ class Calibration(Module):
         self.t0set = self.get("t0set")
         self.calibset = self.get("calibset")
         self.detector = self.get("detector")
+        self.key = self.get("key", default="Hits")
+        self.outkey = self.get("outkey", default="CalibHits")
+        self.key_mc = self.get("key_mc", default="McHits")
+        self.outkey_mc = self.get("outkey_mc", default="CalibMcHits")
         self._pos_dom_channel = None
         self._dir_dom_channel = None
         self._t0_dom_channel = None
@@ -89,9 +106,11 @@ class Calibration(Module):
         else:
             self.log.critical("No detector information loaded.")
 
-    def process(self, blob, key="Hits", outkey="CalibHits"):
+    def process(self, blob):
         if self._should_apply:
-            blob[outkey] = self.apply(blob[key])
+            blob[self.outkey] = self.apply(blob[self.key])
+            if self.key_mc in blob:
+                blob[self.outkey_mc] = self.apply(blob[self.key_mc])
         return blob
 
     def get_detector(self):
@@ -126,6 +145,7 @@ class Calibration(Module):
             # do we ever see McHits here?
             hits = Table.from_template(hits, "Hits")
 
+        is_mc = None
         if hasattr(hits, "dom_id") and hasattr(hits, "channel_id"):
             try:
                 (
@@ -138,10 +158,12 @@ class Calibration(Module):
                     pos_y,
                     pos_z,
                     t0,
+                    pmt_id,
                 ) = _get_calibration_for_hits(hits, self._calib_by_dom_and_channel)
             except KeyError as e:
                 self.log.critical("Wrong calibration (DETX) data provided.")
                 raise
+            is_mc = False
         elif hasattr(hits, "pmt_id"):
             try:
                 (
@@ -154,10 +176,13 @@ class Calibration(Module):
                     pos_y,
                     pos_z,
                     t0,
+                    dom_id,
+                    channel_id,
                 ) = _get_calibration_for_mchits(hits, self._calib_by_pmt_id)
             except KeyError as e:
                 self.log.critical("Wrong calibration (DETX) data provided.")
                 raise
+            is_mc = True
         else:
             raise TypeError(
                 "Don't know how to apply calibration to '{0}'. "
@@ -165,7 +190,7 @@ class Calibration(Module):
                 "'pmt_id'.".format(hits.name)
             )
 
-        if hasattr(hits, "time"):
+        if hasattr(hits, "time") and not is_mc:
             if hits.time.dtype != t0.dtype:
                 time = hits.time.astype("f4") + t0.astype("f4")
                 hits = hits.drop_columns(["time"])
@@ -187,18 +212,28 @@ class Calibration(Module):
             "pos_z": pos_z,
             "t0": t0,
         }
+
+        if is_mc:
+            calib["dom_id"] = dom_id.astype(np.int32)
+            calib["channel_id"] = channel_id.astype(np.int32)
+        else:
+            calib["pmt_id"] = pmt_id.astype(np.int32)
+
         hits_data.update(calib)
-        if correct_slewing and hasattr(hits, "tot"):  # only correct non-MC hits
+
+        if correct_slewing and not is_mc:
             hits_data["time"] -= slew(hits_data["tot"], variant=slewing_variant)
         return Table(
             hits_data, h5loc=hits.h5loc, split_h5=hits.split_h5, name=hits.name
         )
 
     def _create_dom_channel_lookup(self):
-        data = nb.typed.Dict.empty(key_type=nb.types.i8, value_type=nb.types.float64[:, :])
+        data = nb.typed.Dict.empty(
+            key_type=nb.types.i8, value_type=nb.types.float64[:, :]
+        )
         for pmt in self.detector.pmts:
             if pmt.dom_id not in data:
-                data[pmt.dom_id] = np.zeros((31, 9))
+                data[pmt.dom_id] = np.zeros((31, 10))
             data[pmt.dom_id][pmt.channel_id] = np.asarray(
                 [
                     pmt.pos_x,
@@ -210,6 +245,7 @@ class Calibration(Module):
                     pmt.t0,
                     pmt.du,
                     pmt.floor,
+                    pmt.pmt_id,
                 ],
                 dtype=np.float64,
             )
@@ -230,6 +266,8 @@ class Calibration(Module):
                     pmt.t0,
                     pmt.du,
                     pmt.floor,
+                    pmt.dom_id,
+                    pmt.channel_id,
                 ],
                 dtype=np.float64,
             )
@@ -262,7 +300,7 @@ def apply_t0_nb(times, dom_ids, channel_ids, lookup_tables):
 def _get_calibration_for_hits(hits, lookup):
     """Append the position, direction and t0 columns and add t0 to time"""
     n = len(hits)
-    cal = np.empty((n, 9))
+    cal = np.empty((n, 10))
     for i in range(n):
         calib = lookup[hits["dom_id"][i]][hits["channel_id"][i]]
         cal[i] = calib
@@ -274,17 +312,18 @@ def _get_calibration_for_hits(hits, lookup):
     pos_x = cal[:, 0]
     pos_y = cal[:, 1]
     pos_z = cal[:, 2]
+    pmt_id = cal[:, 9]
 
     t0 = cal[:, 6]
 
-    return [dir_x, dir_y, dir_z, du, floor, pos_x, pos_y, pos_z, t0]
+    return [dir_x, dir_y, dir_z, du, floor, pos_x, pos_y, pos_z, t0, pmt_id]
 
 
 @nb.jit
 def _get_calibration_for_mchits(hits, lookup):
     """Append the position, direction and t0 columns and add t0 to time"""
     n_hits = len(hits)
-    cal = np.empty((n_hits, 9))
+    cal = np.empty((n_hits, 11))
     for i in range(n_hits):
         cal[i] = lookup[hits["pmt_id"][i]]
     dir_x = cal[:, 3]
@@ -296,8 +335,10 @@ def _get_calibration_for_mchits(hits, lookup):
     pos_y = cal[:, 1]
     pos_z = cal[:, 2]
     t0 = cal[:, 6]
+    dom_id = cal[:, 9]
+    channel_id = cal[:, 10]
 
-    return [dir_x, dir_y, dir_z, du, floor, pos_x, pos_y, pos_z, t0]
+    return [dir_x, dir_y, dir_z, du, floor, pos_x, pos_y, pos_z, t0, dom_id, channel_id]
 
 
 class CalibrationService(Module):